+++ amqp/src/main/scala/com/colisweb/application/context/amqp/package.scala
@@ -1,0 +1,3 @@
+package com.colisweb.application.context
+
+package object amqp extends AmqpConsumerWithCorrelationId