+++ domain/src/main/scala/com/colisweb/tracing/domain/DomainContext.scala
@@ -1,0 +1,11 @@
+package com.colisweb.tracing.domain
+
+import org.slf4j.Logger
+
+trait DomainContext[F[_]] {
+
+  def addToMDC(key: String, value: String): Unit
+
+  def logger(implicit slf4jLogger: Logger): PureLogger[F]
+
+}