+++ http/client/src/main/scala/com/colisweb/tracing/http/client/package.scala
@@ -1,0 +1,3 @@
+package com.colisweb.tracing.http
+
+package object client extends RequestWithCorrelationId