+++ http/server/src/main/scala/com/colisweb/tracing/http/server/package.scala
@@ -1,0 +1,3 @@
+package com.colisweb.tracing.http
+
+package object server extends HttpTracedEndpointWithExtendedContext with TracedRoutes