+++ http/test/src/test/scala/com/colisweb/tracing/http/test/WrappedCorrelationId.scala
@@ -1,0 +1,3 @@
+package com.colisweb.tracing.http.test
+
+final case class WrappedCorrelationId(correlationId: String)