+++ tapir/src/main/scala/com/colisweb/tracing/tapir/package.scala
@@ -1,3 +1,0 @@
-package com.colisweb.tracing
-
-package object tapir extends TracedRoutes