+++ tracing/src/main/scala/com/colisweb/tracing/context/TracingContextBuilder.scala
@@ -1,0 +1,8 @@
+package com.colisweb.tracing.context
+
+trait TracingContextBuilder[F[_]] {
+  def apply(
+      operationName: String,
+      tags: Tags = Map.empty
+  ): TracingContextResource[F]
+}