+++ tracing/src/main/scala/com/colisweb/tracing/http4s/TracedRequest.scala
@@ -1,0 +1,9 @@
+package com.colisweb.tracing.http4s
+
+import com.colisweb.tracing.context.TracingContext
+import org.http4s.Request
+
+case class TracedRequest[F[_]](
+    request: Request[F],
+    tracingContext: TracingContext[F]
+)